+++ test/integration-test/test-suite/src/test/java/org/apache/shardingsphere/test/integration/engine/BaseITCase.java
@@ -19,18 +19,28 @@ package org.apache.shardingsphere.test.integration.engine;
 
 import lombok.AccessLevel;
 import lombok.Getter;
+import org.apache.shardingsphere.infra.database.DefaultDatabase;
 import org.apache.shardingsphere.infra.database.type.DatabaseType;
 import org.apache.shardingsphere.test.integration.cases.assertion.IntegrationTestCase;
 import org.apache.shardingsphere.test.integration.container.compose.ContainerComposer;
 import org.apache.shardingsphere.test.integration.container.compose.ContainerComposerRegistry;
+import org.apache.shardingsphere.test.integration.env.runtime.scenario.path.ScenarioDataPath;
 import org.apache.shardingsphere.test.integration.framework.param.model.ParameterizedArray;
 import org.apache.shardingsphere.test.integration.framework.runner.ShardingSphereIntegrationTestParameterized;
+import org.h2.tools.RunScript;
 import org.junit.AfterClass;
 import org.junit.Before;
 import org.junit.runner.RunWith;
 
 import javax.sql.DataSource;
+import java.io.FileReader;
+import java.io.IOException;
+import java.sql.Connection;
+import java.sql.SQLException;
+import java.util.Collection;
+import java.util.HashSet;
 import java.util.Map;
+import java.util.Optional;
 import java.util.concurrent.atomic.AtomicInteger;
 
 @RunWith(ShardingSphereIntegrationTestParameterized.class)
@@ -44,6 +54,8 @@ public abstract class BaseITCase {
     private static final int TOTAL_SUITES_COUNT = TotalSuitesCountCalculator.calculate();
     
     private static final AtomicInteger COMPLETED_SUITES_COUNT = new AtomicInteger(0);
+    
+    private static final Collection<String> FILLED_SUITES = new HashSet<>();
     
     private final String mode;
     
@@ -74,11 +86,38 @@ public abstract class BaseITCase {
     }
     
     @Before
-    public void setUp() {
+    public void setUp() throws SQLException, IOException {
         containerComposer.start();
         actualDataSourceMap = containerComposer.getActualDataSourceMap();
         targetDataSource = containerComposer.getTargetDataSource();
         expectedDataSourceMap = containerComposer.getExpectedDataSourceMap();
+        executeLogicDatabaseInitSQLFileOnlyOnce(targetDataSource);
+    }
+    
+    private void executeLogicDatabaseInitSQLFileOnlyOnce(final DataSource targetDataSource) throws SQLException, IOException {
+        if (!FILLED_SUITES.contains(getItKey())) {
+            synchronized (FILLED_SUITES) {
+                if (!FILLED_SUITES.contains(getScenario())) {
+                    executeLogicDatabaseInitSQLFile(targetDataSource);
+                    FILLED_SUITES.add(getItKey());
+                }
+            }
+        }
+    }
+    
+    private void executeLogicDatabaseInitSQLFile(final DataSource targetDataSource) throws SQLException, IOException {
+        Optional<String> logicDatabaseInitSQLFile = new ScenarioDataPath(getScenario()).findActualDatabaseInitSQLFile(DefaultDatabase.LOGIC_NAME, getDatabaseType());
+        if (logicDatabaseInitSQLFile.isPresent()) {
+            executeInitSQL(targetDataSource, logicDatabaseInitSQLFile.get());
+        }
+    }
+    
+    private void executeInitSQL(final DataSource dataSource, final String logicDatabaseInitSQLFile) throws SQLException, IOException {
+        try (
+                Connection connection = dataSource.getConnection();
+                FileReader reader = new FileReader(logicDatabaseInitSQLFile)) {
+            RunScript.execute(connection, reader);
+        }
     }
     
     @AfterClass
